refactor: move dto into controller package
